enum class RenderType 추가 SpriteRenderer 클래스 멤버변수 RenderType 추가

+++ Client/syInput.cpp
@@ -65,8 +65,8 @@ namespace sy
 			GetCursorPos(&MousePos);
 			// 마우스 Pos 를 스크린 좌표에서 특정 클라이언트 좌표로 바꿔준다
 			ScreenToClient(hWnd, &MousePos);
-			mMousePos.x = MousePos.x;
-			mMousePos.y = MousePos.y;
+			mMousePos.x = (float)MousePos.x;
+			mMousePos.y = (float)MousePos.y;
 		}
 		else // 현재 포커스중이 아니다
 		{
